number product in category

+++ src/app/(webpage)/api/products/deleteProduct/[id]/route.ts
@@ -5,6 +5,7 @@ import cloudinary from "@/lib/cloudinary";
 import User from "@/models/User";
 import { getToken } from "next-auth/jwt";
 import Review from "@/models/Review";
+import Category from '@/models/Category';
 
 
 const extractPublicId = (url: string): string => {
@@ -76,7 +77,12 @@ const deleteImagesFromCloudinary = async (urls: string[]) => {
  if (product.images && product.images.length > 0) {
   await deleteImagesFromCloudinary(product.images);
 }
-
+const existingcategory = await Category.findById(product.category);
+if (!existingcategory){
+  return NextResponse.json({ message: 'Error product ' }, { status: 402 });
+}
+  existingcategory.numberproduct -= 1;
+  existingcategory.save();
       // Delete the product from the database
     await Product.findByIdAndDelete(id);
